Revert "Allow setting Tx isolation on Tx start (#384)"
This reverts commit 385682fd0211c76f1187c10708c49da05e65bdcd.
