+++ src/connector/sqlite.rs
@@ -3,7 +3,6 @@ mod error;
 
 pub use rusqlite::{params_from_iter, version as sqlite_version};
 
-use super::IsolationLevel;
 use crate::{
     ast::{Query, Value},
     connector::{metrics, queryable::*, ResultSet},
@@ -165,6 +164,11 @@ impl Queryable for Sqlite {
         self.query_raw(&sql, &params).await
     }
 
+    async fn execute(&self, q: Query<'_>) -> crate::Result<u64> {
+        let (sql, params) = visitor::Sqlite::build(q)?;
+        self.execute_raw(&sql, &params).await
+    }
+
     async fn query_raw(&self, sql: &str, params: &[Value<'_>]) -> crate::Result<ResultSet> {
         metrics::query("sqlite.query_raw", sql, params, move || async move {
             let client = self.client.lock().await;
@@ -187,11 +191,6 @@ impl Queryable for Sqlite {
 
     async fn query_raw_typed(&self, sql: &str, params: &[Value<'_>]) -> crate::Result<ResultSet> {
         self.query_raw(sql, params).await
-    }
-
-    async fn execute(&self, q: Query<'_>) -> crate::Result<u64> {
-        let (sql, params) = visitor::Sqlite::build(q)?;
-        self.execute_raw(&sql, &params).await
     }
 
     async fn execute_raw(&self, sql: &str, params: &[Value<'_>]) -> crate::Result<u64> {
@@ -224,22 +223,6 @@ impl Queryable for Sqlite {
 
     fn is_healthy(&self) -> bool {
         true
-    }
-
-    async fn set_tx_isolation_level(&self, isolation_level: IsolationLevel) -> crate::Result<()> {
-        // SQLite is always "serializable", other modes involve pragmas
-        // and shared cache mode, which is out of scope for now and should be implemented
-        // as part of a separate effort.
-        if !matches!(isolation_level, IsolationLevel::Serializable) {
-            let kind = ErrorKind::invalid_isolation_level(&isolation_level);
-            return Err(Error::builder(kind).build());
-        }
-
-        Ok(())
-    }
-
-    fn requires_isolation_first(&self) -> bool {
-        false
     }
 }
