+++ src/connector/transaction.rs
@@ -1,21 +1,9 @@
 use super::*;
-use crate::{
-    ast::*,
-    error::{Error, ErrorKind},
-};
+use crate::ast::*;
 use async_trait::async_trait;
-use metrics::{decrement_gauge, increment_gauge};
-use std::{fmt, str::FromStr};
 
 extern crate metrics as metrics;
-
-pub(crate) struct TransactionOptions {
-    /// The isolation level to use.
-    pub(crate) isolation_level: Option<IsolationLevel>,
-
-    /// Whether or not to put the isolation level `SET` before or after the `BEGIN`.
-    pub(crate) isolation_first: bool,
-}
+use metrics::{decrement_gauge, increment_gauge};
 
 /// A representation of an SQL database transaction. If not commited, a
 /// transaction will be rolled back by default when dropped.
@@ -27,30 +15,13 @@ pub struct Transaction<'a> {
 }
 
 impl<'a> Transaction<'a> {
-    pub(crate) async fn new(
-        inner: &'a dyn Queryable,
-        begin_stmt: &str,
-        tx_opts: TransactionOptions,
-    ) -> crate::Result<Transaction<'a>> {
+    pub(crate) async fn new(inner: &'a dyn Queryable, begin_stmt: &str) -> crate::Result<Transaction<'a>> {
         let this = Self { inner };
 
-        if tx_opts.isolation_first {
-            if let Some(isolation) = tx_opts.isolation_level {
-                inner.set_tx_isolation_level(isolation).await?;
-            }
-        }
-
         inner.raw_cmd(begin_stmt).await?;
-
-        if !tx_opts.isolation_first {
-            if let Some(isolation) = tx_opts.isolation_level {
-                inner.set_tx_isolation_level(isolation).await?;
-            }
-        }
-
         inner.server_reset_query(&this).await?;
-        increment_gauge!("query_active_transactions", 1.0);
 
+        increment_gauge!("query_active_transactions", 1.0);
         Ok(this)
     }
 
@@ -107,74 +78,5 @@ impl<'a> Queryable for Transaction<'a> {
 
     fn is_healthy(&self) -> bool {
         self.inner.is_healthy()
-    }
-
-    async fn set_tx_isolation_level(&self, isolation_level: IsolationLevel) -> crate::Result<()> {
-        self.inner.set_tx_isolation_level(isolation_level).await
-    }
-
-    fn requires_isolation_first(&self) -> bool {
-        self.inner.requires_isolation_first()
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-/// Controls the locking and row versioning behavior of connections or transactions.
-/// The levels correspond to the ANSI standard isolation levels, plus `Snapshot` for SQL Server.
-///
-/// Details on exact behavior and validity can be found in the documentation of the database vendors:
-/// - [SQL Server documentation].
-/// - [Postgres documentation].
-/// - [MySQL documentation].
-/// - [SQLite documentation].
-///
-/// [SQL Server documentation]: https://docs.microsoft.com/en-us/sql/t-sql/statements/set-transaction-isolation-level-transact-sql?view=sql-server-ver15
-/// [Postgres documentation]: https://www.postgresql.org/docs/current/sql-set-transaction.html
-/// [MySQL documentation]: https://dev.mysql.com/doc/refman/8.0/en/innodb-transaction-isolation-levels.html
-/// [SQLite documentation]: https://www.sqlite.org/isolation.html
-///
-pub enum IsolationLevel {
-    ReadUncommitted,
-    ReadCommitted,
-    RepeatableRead,
-    Snapshot,
-    Serializable,
-}
-
-impl fmt::Display for IsolationLevel {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::ReadUncommitted => write!(f, "READ UNCOMMITTED"),
-            Self::ReadCommitted => write!(f, "READ COMMITTED"),
-            Self::RepeatableRead => write!(f, "REPEATABLE READ"),
-            Self::Snapshot => write!(f, "SNAPSHOT"),
-            Self::Serializable => write!(f, "SERIALIZABLE"),
-        }
-    }
-}
-
-impl FromStr for IsolationLevel {
-    type Err = Error;
-
-    fn from_str(s: &str) -> crate::Result<Self> {
-        match s {
-            "READ UNCOMMITTED" => Ok(Self::ReadUncommitted),
-            "READ COMMITTED" => Ok(Self::ReadCommitted),
-            "REPEATABLE READ" => Ok(Self::RepeatableRead),
-            "SNAPSHOT" => Ok(Self::Snapshot),
-            "SERIALIZABLE" => Ok(Self::Serializable),
-            _ => {
-                let kind = ErrorKind::conversion(format!("Invalid isolation level `{}`", s));
-                Err(Error::builder(kind).build())
-            }
-        }
-    }
-}
-impl TransactionOptions {
-    pub(crate) fn new(isolation_level: Option<IsolationLevel>, isolation_first: bool) -> Self {
-        Self {
-            isolation_level,
-            isolation_first,
-        }
     }
 }
