+++ src/error.rs
@@ -1,5 +1,4 @@
 //! Error module
-use crate::connector::IsolationLevel;
 use std::{borrow::Cow, fmt, io, num};
 use thiserror::Error;
 
@@ -244,9 +243,6 @@ pub enum ErrorKind {
     #[error("Transaction was already closed: {}", _0)]
     TransactionAlreadyClosed(String),
 
-    #[error("Invalid isolation level: {}", _0)]
-    InvalidIsolationLevel(String),
-
     #[error("Error creating UUID, {}", _0)]
     UUIDError(String),
 
@@ -279,10 +275,6 @@ impl ErrorKind {
             in_use,
             timeout: timeout.as_secs(),
         }
-    }
-
-    pub(crate) fn invalid_isolation_level(isolation_level: &IsolationLevel) -> Self {
-        Self::InvalidIsolationLevel(isolation_level.to_string())
     }
 }
